fix issue 119
if you focus an atom settings tab and try to open php-debug, atom.workspace.getActivePaneItem()
will return a SettingsView object instead of a TextEditor object.
TextEditor objects have the gutterWithName method, SettingsView objects don't.
This fix merely makes the error disappear.
In my opinion, a proper fix would be clean code, lot's of rewriting...
This function is related to breakpoints and as such, it should be implemented somewhere in
lib/breakpoint/*.coffee.
Now it's just messy and buggy and polluting the .mainModule. namespace.